refactor(desktop): decompose PromptGroup.tsx into utils, hooks, and components
Extract pure functions, hooks, and inline components from the 1457-line
PromptGroup.tsx into co-located modules per project conventions:

- utils/convertFiles: convertBlobUrlToDataUrl, convertPromptInputFiles
- utils/fetchGitHubIssueFiles: GitHub issue content fetching + sanitization
- hooks/useBranchData: branch/worktree queries and derived memos
- hooks/useLinkedItems: linked issue/PR handlers
- components/AttachmentButtons, ProjectPickerPill, CompareBaseBranchPickerInline
- constants.ts: shared PILL_BUTTON_CLASS

Remove unnecessary useCallbacks from handlers only used by non-memoized
children. Keep useCallback on handleCreate (dependency of Cmd+Enter effect).
