+++ apps/desktop/src/renderer/components/NewWorkspaceModal/components/PromptGroup/components/AttachmentButtons/index.ts
@@ -1,0 +1,1 @@
+export { AttachmentButtons } from "./AttachmentButtons";
