+++ apps/desktop/src/renderer/components/NewWorkspaceModal/components/PromptGroup/components/ProjectPickerPill/ProjectPickerPill.tsx
@@ -1,0 +1,125 @@
+import { PromptInputButton } from "@superset/ui/ai-elements/prompt-input";
+import {
+	Command,
+	CommandEmpty,
+	CommandGroup,
+	CommandInput,
+	CommandItem,
+	CommandList,
+	CommandSeparator,
+} from "@superset/ui/command";
+import { Popover, PopoverContent, PopoverTrigger } from "@superset/ui/popover";
+import { useState } from "react";
+import { HiCheck, HiChevronUpDown } from "react-icons/hi2";
+import { LuFolderGit, LuFolderOpen } from "react-icons/lu";
+import { ProjectThumbnail } from "renderer/screens/main/components/WorkspaceSidebar/ProjectSection/ProjectThumbnail";
+import { PILL_BUTTON_CLASS } from "../../constants";
+
+interface ProjectOption {
+	id: string;
+	name: string;
+	color: string;
+	githubOwner: string | null;
+	iconUrl: string | null;
+	hideImage: boolean | null;
+}
+
+interface ProjectPickerPillProps {
+	selectedProject: ProjectOption | undefined;
+	recentProjects: ProjectOption[];
+	onSelectProject: (projectId: string) => void;
+	onImportRepo: () => void;
+	onNewProject: () => void;
+}
+
+export function ProjectPickerPill({
+	selectedProject,
+	recentProjects,
+	onSelectProject,
+	onImportRepo,
+	onNewProject,
+}: ProjectPickerPillProps) {
+	const [open, setOpen] = useState(false);
+
+	return (
+		<Popover open={open} onOpenChange={setOpen}>
+			<PopoverTrigger asChild>
+				<PromptInputButton
+					className={`${PILL_BUTTON_CLASS} px-1.5 gap-1 text-foreground w-auto max-w-[140px]`}
+				>
+					{selectedProject && (
+						<ProjectThumbnail
+							projectId={selectedProject.id}
+							projectName={selectedProject.name}
+							projectColor={selectedProject.color}
+							githubOwner={selectedProject.githubOwner}
+							iconUrl={selectedProject.iconUrl}
+							hideImage={selectedProject.hideImage ?? false}
+							className="!size-3"
+						/>
+					)}
+					<span className="truncate">
+						{selectedProject?.name ?? "Select project"}
+					</span>
+					<HiChevronUpDown className="size-3 shrink-0 text-muted-foreground" />
+				</PromptInputButton>
+			</PopoverTrigger>
+			<PopoverContent align="start" className="w-60 p-0">
+				<Command>
+					<CommandInput placeholder="Search projects..." />
+					<CommandList>
+						<CommandEmpty>No projects found.</CommandEmpty>
+						<CommandGroup>
+							{recentProjects.map((project) => (
+								<CommandItem
+									key={project.id}
+									value={project.name}
+									onSelect={() => {
+										onSelectProject(project.id);
+										setOpen(false);
+									}}
+								>
+									<ProjectThumbnail
+										projectId={project.id}
+										projectName={project.name}
+										projectColor={project.color}
+										githubOwner={project.githubOwner}
+										iconUrl={project.iconUrl}
+										hideImage={project.hideImage ?? false}
+									/>
+									{project.name}
+									{project.id === selectedProject?.id && (
+										<HiCheck className="ml-auto size-4" />
+									)}
+								</CommandItem>
+							))}
+						</CommandGroup>
+						<CommandSeparator alwaysRender />
+						<CommandGroup forceMount>
+							<CommandItem
+								forceMount
+								onSelect={() => {
+									setOpen(false);
+									onImportRepo();
+								}}
+							>
+								<LuFolderOpen className="size-4" />
+								Open project
+							</CommandItem>
+							<CommandItem
+								forceMount
+								onSelect={() => {
+									setOpen(false);
+									onNewProject();
+								}}
+							>
+								<LuFolderGit className="size-4" />
+								New project
+							</CommandItem>
+						</CommandGroup>
+					</CommandList>
+				</Command>
+			</PopoverContent>
+		</Popover>
+	);
+}
