+++ apps/desktop/src/renderer/components/NewWorkspaceModal/components/PromptGroup/components/ProjectPickerPill/index.ts
@@ -1,0 +1,1 @@
+export { ProjectPickerPill } from "./ProjectPickerPill";
