+++ apps/desktop/src/renderer/components/NewWorkspaceModal/components/PromptGroup/constants.ts
@@ -1,0 +1,2 @@
+export const PILL_BUTTON_CLASS =
+	"!h-[22px] min-h-0 rounded-md border-[0.5px] border-border bg-foreground/[0.04] shadow-none text-[11px]";
