+++ apps/desktop/src/renderer/components/NewWorkspaceModal/components/PromptGroup/hooks/useBranchData/index.ts
@@ -1,0 +1,1 @@
+export { useBranchData } from "./useBranchData";
