+++ apps/desktop/src/renderer/components/NewWorkspaceModal/components/PromptGroup/hooks/useLinkedItems/index.ts
@@ -1,0 +1,1 @@
+export { useLinkedItems } from "./useLinkedItems";
