+++ apps/desktop/src/renderer/components/NewWorkspaceModal/components/PromptGroup/utils/convertFiles/index.ts
@@ -1,0 +1,5 @@
+export type { ConvertedFile } from "./convertFiles";
+export {
+	convertBlobUrlToDataUrl,
+	convertPromptInputFiles,
+} from "./convertFiles";
