+++ apps/desktop/src/renderer/components/NewWorkspaceModal/components/PromptGroup/utils/fetchGitHubIssueFiles/index.ts
@@ -1,0 +1,1 @@
+export { fetchGitHubIssueFiles } from "./fetchGitHubIssueFiles";
